9.4.3 Verify VerseFake plays the verse template role
This test prevents the unfortunate circumstance in which tests that
collaborate with fakes sometimes continue to pass even after changes to
production code have broken the application. If, for example, the API of
the verse template role changes from `lyrics` to `sing`, a test will now
fail unless you update every role player and the `testPlaysVerseRole` to
reflect the new interface. This forces all players of a role, even your
fakes, to stay in sync with the API